+++ br/pkg/task/common.go
@@ -20,7 +20,6 @@ import (
 	backuppb "github.com/pingcap/kvproto/pkg/brpb"
 	"github.com/pingcap/kvproto/pkg/encryptionpb"
 	"github.com/pingcap/log"
-	filter "github.com/pingcap/tidb-tools/pkg/table-filter"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
@@ -30,7 +29,6 @@ import (
 	"github.com/tikv/migration/br/pkg/metautil"
 	"github.com/tikv/migration/br/pkg/storage"
 	pd "github.com/tikv/pd/client"
-	"go.etcd.io/etcd/pkg/transport"
 	"go.uber.org/zap"
 	"google.golang.org/grpc/keepalive"
 )
@@ -80,84 +78,6 @@ const (
 	crypterAES192KeyLen = 24
 	crypterAES256KeyLen = 32
 )
-
-// TLSConfig is the common configuration for TLS connection.
-type TLSConfig struct {
-	CA   string `json:"ca" toml:"ca"`
-	Cert string `json:"cert" toml:"cert"`
-	Key  string `json:"key" toml:"key"`
-}
-
-// IsEnabled checks if TLS open or not.
-func (tls *TLSConfig) IsEnabled() bool {
-	return tls.CA != ""
-}
-
-// ToTLSConfig generate tls.Config.
-func (tls *TLSConfig) ToTLSConfig() (*tls.Config, error) {
-	tlsInfo := transport.TLSInfo{
-		CertFile:      tls.Cert,
-		KeyFile:       tls.Key,
-		TrustedCAFile: tls.CA,
-	}
-	tlsConfig, err := tlsInfo.ClientConfig()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	return tlsConfig, nil
-}
-
-// ParseFromFlags parses the TLS config from the flag set.
-func (tls *TLSConfig) ParseFromFlags(flags *pflag.FlagSet) (err error) {
-	tls.CA, tls.Cert, tls.Key, err = ParseTLSTripleFromFlags(flags)
-	return
-}
-
-// Config is the common configuration for all BRIE tasks.
-type Config struct {
-	storage.BackendOptions
-
-	Storage             string    `json:"storage" toml:"storage"`
-	PD                  []string  `json:"pd" toml:"pd"`
-	TLS                 TLSConfig `json:"tls" toml:"tls"`
-	RateLimit           uint64    `json:"rate-limit" toml:"rate-limit"`
-	ChecksumConcurrency uint      `json:"checksum-concurrency" toml:"checksum-concurrency"`
-	Concurrency         uint32    `json:"concurrency" toml:"concurrency"`
-	Checksum            bool      `json:"checksum" toml:"checksum"`
-	SendCreds           bool      `json:"send-credentials-to-tikv" toml:"send-credentials-to-tikv"`
-	// LogProgress is true means the progress bar is printed to the log instead of stdout.
-	LogProgress bool `json:"log-progress" toml:"log-progress"`
-
-	// CaseSensitive should not be used.
-	//
-	// Deprecated: This field is kept only to satisfy the cyclic dependency with TiDB. This field
-	// should be removed after TiDB upgrades the BR dependency.
-	CaseSensitive bool
-
-	// NoCreds means don't try to load cloud credentials
-	NoCreds bool `json:"no-credentials" toml:"no-credentials"`
-
-	CheckRequirements bool `json:"check-requirements" toml:"check-requirements"`
-	// EnableOpenTracing is whether to enable opentracing
-	EnableOpenTracing bool `json:"enable-opentracing" toml:"enable-opentracing"`
-	// SkipCheckPath skips verifying the path
-	// deprecated
-	SkipCheckPath bool `json:"skip-check-path" toml:"skip-check-path"`
-	// Filter should not be used, use TableFilter instead.
-	//
-	// Deprecated: This field is kept only to satisfy the cyclic dependency with TiDB. This field
-	// should be removed after TiDB upgrades the BR dependency.
-	Filter filter.MySQLReplicationRules
-
-	SwitchModeInterval time.Duration `json:"switch-mode-interval" toml:"switch-mode-interval"`
-
-	// GrpcKeepaliveTime is the interval of pinging the server.
-	GRPCKeepaliveTime time.Duration `json:"grpc-keepalive-time" toml:"grpc-keepalive-time"`
-	// GrpcKeepaliveTimeout is the max time a grpc conn can keep idel before killed.
-	GRPCKeepaliveTimeout time.Duration `json:"grpc-keepalive-timeout" toml:"grpc-keepalive-timeout"`
-
-	CipherInfo backuppb.CipherInfo `json:"-" toml:"-"`
-}
 
 // DefineCommonFlags defines the flags common to all BRIE commands.
 func DefineCommonFlags(flags *pflag.FlagSet) {
@@ -213,23 +133,6 @@ func DefineCommonFlags(flags *pflag.FlagSet) {
 	flags.String(flagCipherKeyFile, "", "FilePath, its content is used as the cipher-key")
 
 	storage.DefineFlags(flags)
-}
-
-// ParseTLSTripleFromFlags parses the (ca, cert, key) triple from flags.
-func ParseTLSTripleFromFlags(flags *pflag.FlagSet) (ca, cert, key string, err error) {
-	ca, err = flags.GetString(flagCA)
-	if err != nil {
-		return
-	}
-	cert, err = flags.GetString(flagCert)
-	if err != nil {
-		return
-	}
-	key, err = flags.GetString(flagKey)
-	if err != nil {
-		return
-	}
-	return
 }
 
 func parseCipherType(t string) (encryptionpb.EncryptionMethod, error) {
@@ -291,139 +194,6 @@ func checkCipherKeyMatch(cipher *backuppb.CipherInfo) bool {
 	default:
 		return false
 	}
-}
-
-func (cfg *Config) parseCipherInfo(flags *pflag.FlagSet) error {
-	crypterStr, err := flags.GetString(flagCipherType)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	cfg.CipherInfo.CipherType, err = parseCipherType(crypterStr)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	if cfg.CipherInfo.CipherType == encryptionpb.EncryptionMethod_PLAINTEXT {
-		return nil
-	}
-
-	key, err := flags.GetString(flagCipherKey)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	keyFilePath, err := flags.GetString(flagCipherKeyFile)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	cfg.CipherInfo.CipherKey, err = getCipherKeyContent(key, keyFilePath)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	if !checkCipherKeyMatch(&cfg.CipherInfo) {
-		return errors.Annotate(berrors.ErrInvalidArgument, "crypter method and key length not match")
-	}
-
-	return nil
-}
-
-func (cfg *Config) normalizePDURLs() error {
-	for i := range cfg.PD {
-		var err error
-		cfg.PD[i], err = normalizePDURL(cfg.PD[i], cfg.TLS.IsEnabled())
-		if err != nil {
-			return errors.Trace(err)
-		}
-	}
-	return nil
-}
-
-// ParseFromFlags parses the config from the flag set.
-func (cfg *Config) ParseFromFlags(flags *pflag.FlagSet) error {
-	var err error
-	if cfg.Storage, err = flags.GetString(flagStorage); err != nil {
-		return errors.Trace(err)
-	}
-	if cfg.SendCreds, err = flags.GetBool(flagSendCreds); err != nil {
-		return errors.Trace(err)
-	}
-	if cfg.NoCreds, err = flags.GetBool(flagNoCreds); err != nil {
-		return errors.Trace(err)
-	}
-	if cfg.Concurrency, err = flags.GetUint32(flagConcurrency); err != nil {
-		return errors.Trace(err)
-	}
-	if cfg.Checksum, err = flags.GetBool(flagChecksum); err != nil {
-		return errors.Trace(err)
-	}
-	if cfg.ChecksumConcurrency, err = flags.GetUint(flagChecksumConcurrency); err != nil {
-		return errors.Trace(err)
-	}
-
-	var rateLimit, rateLimitUnit uint64
-	if rateLimit, err = flags.GetUint64(flagRateLimit); err != nil {
-		return errors.Trace(err)
-	}
-	if rateLimitUnit, err = flags.GetUint64(flagRateLimitUnit); err != nil {
-		return errors.Trace(err)
-	}
-	cfg.RateLimit = rateLimit * rateLimitUnit
-
-	checkRequirements, err := flags.GetBool(flagCheckRequirement)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	cfg.CheckRequirements = checkRequirements
-
-	cfg.SwitchModeInterval, err = flags.GetDuration(flagSwitchModeInterval)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	cfg.GRPCKeepaliveTime, err = flags.GetDuration(flagGrpcKeepaliveTime)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	cfg.GRPCKeepaliveTimeout, err = flags.GetDuration(flagGrpcKeepaliveTimeout)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	cfg.EnableOpenTracing, err = flags.GetBool(flagEnableOpenTracing)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	if cfg.SwitchModeInterval <= 0 {
-		return errors.Annotatef(berrors.ErrInvalidArgument, "--switch-mode-interval must be positive, %s is not allowed", cfg.SwitchModeInterval)
-	}
-
-	if err = cfg.BackendOptions.ParseFromFlags(flags); err != nil {
-		return errors.Trace(err)
-	}
-	if err = cfg.TLS.ParseFromFlags(flags); err != nil {
-		return errors.Trace(err)
-	}
-	cfg.PD, err = flags.GetStringSlice(flagPD)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if len(cfg.PD) == 0 {
-		return errors.Annotate(berrors.ErrInvalidArgument, "must provide at least one PD server address")
-	}
-	if cfg.SkipCheckPath, err = flags.GetBool(flagSkipCheckPath); err != nil {
-		return errors.Trace(err)
-	}
-	if cfg.SkipCheckPath {
-		log.L().Info("--skip-check-path is deprecated, need explicitly set it anymore")
-	}
-
-	if err = cfg.parseCipherInfo(flags); err != nil {
-		return errors.Trace(err)
-	}
-
-	return cfg.normalizePDURLs()
 }
 
 // NewMgr creates a new mgr at the given PD address.
@@ -566,20 +336,6 @@ func GetKeepalive(cfg *Config) keepalive.ClientParameters {
 	return keepalive.ClientParameters{
 		Time:    cfg.GRPCKeepaliveTime,
 		Timeout: cfg.GRPCKeepaliveTimeout,
-	}
-}
-
-// adjust adjusts the abnormal config value in the current config.
-// useful when not starting BR from CLI (e.g. from BRIE in SQL).
-func (cfg *Config) adjust() {
-	if cfg.GRPCKeepaliveTime == 0 {
-		cfg.GRPCKeepaliveTime = defaultGRPCKeepaliveTime
-	}
-	if cfg.GRPCKeepaliveTimeout == 0 {
-		cfg.GRPCKeepaliveTimeout = defaultGRPCKeepaliveTimeout
-	}
-	if cfg.ChecksumConcurrency == 0 {
-		cfg.ChecksumConcurrency = variable.DefChecksumTableConcurrency
 	}
 }
