+++ br/pkg/task/restore_raw.go
@@ -8,7 +8,6 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 	berrors "github.com/tikv/migration/br/pkg/errors"
 	"github.com/tikv/migration/br/pkg/glue"
 	"github.com/tikv/migration/br/pkg/metautil"
@@ -16,43 +15,13 @@ import (
 	"github.com/tikv/migration/br/pkg/summary"
 )
 
-// RestoreRawConfig is the configuration specific for raw kv restore tasks.
-type RestoreRawConfig struct {
-	RawKvConfig
-	RestoreCommonConfig
-}
-
 // DefineRawRestoreFlags defines common flags for the backup command.
 func DefineRawRestoreFlags(command *cobra.Command) {
 	command.Flags().StringP(flagKeyFormat, "", "hex", "start/end key format, support raw|escaped|hex")
-	command.Flags().StringP(flagTiKVColumnFamily, "", "default", "restore specify cf, correspond to tikv cf")
 	command.Flags().StringP(flagStartKey, "", "", "restore raw kv start key, key is inclusive")
 	command.Flags().StringP(flagEndKey, "", "", "restore raw kv end key, key is exclusive")
 
 	DefineRestoreCommonFlags(command.PersistentFlags())
-}
-
-// ParseFromFlags parses the backup-related flags from the flag set.
-func (cfg *RestoreRawConfig) ParseFromFlags(flags *pflag.FlagSet) error {
-	var err error
-	cfg.Online, err = flags.GetBool(flagOnline)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	err = cfg.RestoreCommonConfig.ParseFromFlags(flags)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	return cfg.RawKvConfig.ParseFromFlags(flags)
-}
-
-func (cfg *RestoreRawConfig) adjust() {
-	cfg.Config.adjust()
-	cfg.RestoreCommonConfig.adjust()
-
-	if cfg.Concurrency == 0 {
-		cfg.Concurrency = defaultRestoreConcurrency
-	}
 }
 
 // RunRestoreRaw starts a raw kv restore task inside the current goroutine.
@@ -99,7 +68,7 @@ func RunRestoreRaw(c context.Context, g glue.Glue, cmdName string, cfg *RestoreR
 		return errors.Annotate(berrors.ErrRestoreModeMismatch, "cannot do raw restore from transactional data")
 	}
 
-	files, err := client.GetFilesInRawRange(cfg.StartKey, cfg.EndKey, cfg.CF)
+	files, err := client.GetFilesInRawRange(cfg.StartKey, cfg.EndKey, "default")
 	if err != nil {
 		return errors.Trace(err)
 	}
