+++ br/tests/rawkv/run.py
@@ -16,16 +16,38 @@ class rawkvTester:
         self.br_storage = global_args.br_storage
 
 
-    def test_rawkv(self):
+    def test_dst_apiv1(self):
+        self._run_br_test("v1")
+        self._success_msg(test_dst_apiv1.__name__)
+
+
+    def test_dst_apiv1ttl(self):
+        self._run_br_test("v1ttl")
+        self._success_msg(test_dst_apiv1ttl.__name__)
+
+
+    def test_dst_apiv2(self):
+        self._run_br_test("v2")
+        self._success_msg(test_dst_apiv2.__name__)
+
+
+    def _success_msg(self, case_name):
+        print("PASSED: {}".forma(case_name))
+
+
+    def _run_br_test(self, dst_api_version):
         outer_start, outer_end = "31", "3130303030303030"
         inner_start, inner_end = "311111", "311122"
+
+        # clean the data range to be tested
+        self._clean_range(outer_start, outer_end)
         cs_outer_empty = self._get_checksum(outer_start, outer_end)
 
         # prepare and backup data
         self._randgen(outer_start, outer_end)
         self._run_cmd(self.helper, "-pd", self.pd, "-mode", "put",
                 "-put-data", "311121:31, 31112100:32, 311122:33, 31112200:34, 3111220000:35, 311123:36")
-        self._backup_range(outer_start, outer_end)
+        self._backup_range(outer_start, outer_end, dst_api_version)
         cs_outer_origin = self._get_checksum(outer_start, outer_end)
         cs_inner_origin = self._get_checksum(inner_start, inner_end)
 
@@ -33,6 +55,7 @@ class rawkvTester:
         self._clean_range(outer_start, outer_end)
         cs_outer_clean = self._get_checksum(outer_start, outer_end)
         self._assert("clean range failed, checksum mismatch.\n  actual: {}\n  expect: {}", cs_outer_clean, cs_outer_empty)
+        print("outer_start: {}, outer_end: {}".format(outer_start, outer_end))
         self._restore_range(outer_start, outer_end)
         cs_outer_restore = self._get_checksum(outer_start, outer_end)
         self._assert("restore failed, checksum mismatch.\n  actual: {}\n  expect: {}", cs_outer_restore, cs_outer_origin)
@@ -46,9 +69,10 @@ class rawkvTester:
         self._assert("restore failed, checksum mismatch.\n  actual: {}\n  expect: {}", cs_inner_restore, cs_inner_origin)
 
 
-    def _backup_range(self, start_key, end_key):
+    def _backup_range(self, start_key, end_key, dst_api_version):
         self._run_cmd(self.br, "--pd", self.pd, "backup", "raw", "-s", self.br_storage,
-                "--start", start_key, "--end", end_key, "--format", "hex", "--check-requirements=false")
+                "--start", start_key, "--end", end_key, "--format", "hex", "--dst-api-version", dst_api_version,
+                "--check-requirements=false")
 
 
     def _restore_range(self, start_key, end_key):
@@ -70,7 +94,7 @@ class rawkvTester:
         if matched:
             return str(matched.group(0))[len("Checksum result: "):]
         else:
-            self._exit_with_error("get checksum failed:\n  start_key: {}\n  end_key: {}", start_key, end_key)
+            self._exit_with_error("get checksum failed:\n  start_key: {}\n  end_key: {}".format(start_key, end_key))
 
 
     def _run_cmd(self, cmd, *args):
@@ -79,7 +103,12 @@ class rawkvTester:
         for arg in args:
             cmd_list.append(arg)
 
-        output = subprocess.check_output(cmd_list, stderr=sys.stderr, universal_newlines=True)
+        # CalledProcessError
+        try:
+            output = subprocess.run(cmd_list, universal_newlines=True, check=True, stdout=subprocess.PIPE).stdout
+        except  subprocess.CalledProcessError as e:
+            self._exit_with_error("run command failed:\n  cmd: {}\n  stdout: {}\n  stderr: {}".format(e.cmd, e.stdout, e.stderr))
+
         return str(output)
 
 
@@ -100,7 +129,9 @@ class rawkvTester:
 def main():
     args = parse_args()
     tester = rawkvTester(args)
-    tester.test_rawkv()
+    tester.test_dst_apiv1()
+    tester.test_dst_apiv1ttl()
+    tester.test_dst_apiv2()
 
 
 def parse_args():
